Refactor(Multi-Step Form): refactors Form steps into separate component files and adds config
