+++ app/components/Steps/StepThree.tsx
@@ -1,0 +1,5 @@
+import { DisplayWrapper } from "../DisplayWrapper/DisplayWrapper";
+
+export const StepThree = () => {
+  return <DisplayWrapper />;
+};
